Tests for models modified: - all usernames start with the word "test" for better handling image deletion once the tests are completed - base setup for models created separately.
